LRCI-1447 maxLogSize should be a long

+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/JenkinsResultsParserUtil.java
@@ -495,7 +495,7 @@ public class JenkinsResultsParserUtil {
 
 	public static void executeBashService(
 		final String command, final File baseDir,
-		final Map<String, String> environments, final int maxLogSize) {
+		final Map<String, String> environments, final long maxLogSize) {
 
 		Runnable runnable = new Runnable() {
